使事件系统 改成自己写的EventCenter.cs
# Conflicts:
#	Assets/GameScript/AOT/PatchLogic/HFView.cs

+++ Assets/GameScript/HotUpdate/Logic/Bag/BagMainView.cs
@@ -17,7 +17,7 @@ namespace Bag
             base.OnInit();
 
             _closeButton.onClick.Set(OnClickCloseMainView);
-            EventCenter.Instance.Bind<string>(EventEnum.EE_test, OnEventTest);
+            EventCenter.Instance.Bind<string>((int)EventEnum.EE_test, OnEventTest);
 
             _propList.itemRenderer = OnRenderPropList;
             _propList.onClickItem.Add(OnClickItemPropList);
@@ -110,7 +110,7 @@ namespace Bag
         public override void Dispose()
         {
             mSelectItemDto = null;
-            EventCenter.Instance.UnBind<string>(EventEnum.EE_test, OnEventTest);
+            EventCenter.Instance.UnBind<string>((int)EventEnum.EE_test, OnEventTest);
             base.Dispose();
         }
